Rename mapping to config
